+++ src/main/java/com/kllhy/roadmap/roadmap/domain/event/enums/ActiveStatus.java
@@ -1,0 +1,5 @@
+package com.kllhy.roadmap.roadmap.domain.event.enums;
+
+public enum ActiveStatus {
+    ACTIVE, INACTIVE
+}
