+++ src/main/java/com/kllhy/roadmap/roadmap/domain/event/enums/EventType.java
@@ -1,0 +1,5 @@
+package com.kllhy.roadmap.roadmap.domain.event.enums;
+
+public enum EventType {
+    CREATED, UPDATED, DELETED
+}
